requirements.txt
streamlit
numpy
pillow
opencv-python-headless
scikit-image>=0.22
scipy
